code-clash-backend/requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.7
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
uritemplate==3.0.0
urllib3==1.23
